Moved internal tests to seperate file
